src/main/java/com/example/springbootjunitmockito/service/StudentServiceImpl.java
package com.example.springbootjunitmockito.service;

import com.example.springbootjunitmockito.exception.BadRequestException;
import com.example.springbootjunitmockito.exception.StudentNotFoundException;
import com.example.springbootjunitmockito.model.Student;
import com.example.springbootjunitmockito.repository.StudentRepository;
import lombok.AllArgsConstructor;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@AllArgsConstructor
public class StudentServiceImpl implements StudentService {

    @Autowired
    StudentRepository studentRepository;

    @Override
    public List<Student> getAllStudents() {
        return studentRepository.findAll();
    }

    @Override
    public void addStudent(Student student) {
        Boolean existsEmail = studentRepository.selectExistsEmail(student.getEmail());
        if (existsEmail) {
            throw new BadRequestException("Email" + student.getEmail() + "taken");
        }
        studentRepository.save(student);
    }

    @Override
    public void deleteStudent(Long studentId) {
        if (!studentRepository.existsById(studentId)) {
            throw new StudentNotFoundException("Student with id " + studentId + " does not exists");
        }
        studentRepository.deleteById(studentId);
    }
}
